ref: split oidc.service into multiple files
